feat: export repository config contract
Compose AppManifest from the reusable RepositoryConfig slice.\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
