+++ .changeset/repository-config.md
@@ -1,0 +1,5 @@
+---
+'@ankhorage/contracts': minor
+---
+
+Export the reusable `RepositoryConfig` contract slice and compose `AppManifest` from it.
